implemented tower upgrade

+++ TowerDefenceGame/Assets/Scripts/DefaultUnitModel.cs
@@ -2,7 +2,7 @@
 
 public class DefaultUnitModel : UnitModel
 {
-    public DefaultUnitModel(float health, float damage, float gold, float speed) : base(health, damage, gold, speed)
+    public DefaultUnitModel(float health, float damage, int gold, float speed) : base(health, damage, gold, speed)
     {
     }
 }
